pyproject.toml
[project]
name = "invoice-multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "langgraph>=1.0.7",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.52.2",
]
